Move config files to config/
Also update lint-staged to 10 RC
