+++ package.json
@@ -15,10 +15,10 @@
     "build": "yarn clean && gatsby build",
     "clean": "gatsby clean",
     "develop": "gatsby develop",
-    "format": "prettier \"**/*.{css,gql,graphql,html,json,md,mdx,sass,scss,toml,yaml,yml}\" --write",
+    "format": "prettier --write --config config/.prettierrc.js \"**/*.{css,gql,graphql,html,json,md,mdx,sass,scss,toml,yaml,yml}\"",
     "info": "gatsby info",
     "info:copy": "yarn info --clipboard",
-    "lint": "eslint . --ext js,jsx,ts,tsx --fix",
+    "lint": "eslint --fix --ext js,jsx,ts,tsx --config config/.eslintrc.js .",
     "repl": "gatsby repl",
     "serve": "gatsby serve",
     "start": "yarn develop",
@@ -55,7 +55,7 @@
     "eslint-plugin-react": "^7.16.0",
     "eslint-plugin-standard": "^4.0.1",
     "husky": "^3.0.9",
-    "lint-staged": "^9.4.2",
+    "lint-staged": "^10.0.0-1",
     "prettier": "^1.18.2",
     "typescript": "^3.7.0-dev.20191021",
     "webpack": "^4.41.2"
